Dave/comment reply

+++ backend/controllers/forum-controller.js
@@ -2,23 +2,19 @@ const HttpError = require("../models/http-error");
 const { validationResult } = require("express-validator");
 const { v4: uuidv4 } = require("uuid");
 const Forum = require("../models/forum");
+const Comment = require("../models/comment");
 
 // Create forum
 const createForum = async (req, res, next) => {
-  // const errors = validationResult(req);
-  // if (!errors.isEmpty()) {
-  //   return next(
-  //     new HttpError('Invalid inputs passed, please check your data.', 422)
-  //   );
-  // }
 
-  const { creator, headline, topic, initComment } = req.body;
+  const { creator, headline, topic, initComment, anon } = req.body;
 
   const createdForum = new Forum({
     creator,
     headline,
     topic,
     initComment,
+    anon
   });
 
   try {
@@ -158,7 +154,7 @@ const updateForum = async (req, res, next) => {
     throw new HttpError("Invalid inputs passed", 422);
   }
 
-  const { headline, topic, initComment } = req.body;
+  const { headline, topic, initComment, rating, anon } = req.body;
   const forumId = req.params.fid;
 
   let forum;
@@ -175,6 +171,8 @@ const updateForum = async (req, res, next) => {
   forum.headline = headline;
   forum.topic =  topic;
   forum.initComment = initComment;
+  forum.rating = rating;
+  forum.anon = anon;
 
   try {
     await forum.save();
@@ -193,28 +191,21 @@ const updateForum = async (req, res, next) => {
 const deleteForum = async (req, res, next) => {
   const forumId = req.params.fid;
 
-  let forum;
   try {
-    forum = await Forum.findById(forumId);
-  } catch (err) {
-    const error = new HttpError(
-      "Delete - Something went wrong, could not find a forum",
-      500
-    );
-    return next(error);
-  }
-
-  try {
+    await Comment.deleteMany({ forum: forumId });
+    const forum = await Forum.findById(forumId);
+    if (!forum) {
+      throw new HttpError("forum not found", 404);
+    }
     await forum.deleteOne();
+    res.status(200).json({ message: "Deleted forum and associated comments" });
   } catch (err) {
     const error = new HttpError(
-      "Something went wrong, could not delete forum",
+      "Error with deleting comments",
       500
     );
     return next(error);
   }
-
-  res.status(200).json({ message: "Deleted forum" });
 };
 
 exports.createForum = createForum;
